requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pandas==2.1.0
jinja2==3.1.2
matplotlib==3.8.0
mdutils==1.6.0
pdfkit==1.0.0
requests==2.31.0
argparse==1.4.0
python-dateutil==2.8.2
colorama==0.4.6